add endpoints-framework-auth artifact
Rather than shipping EspAuthenticator with the core framework
artifact, we create a new "bridge" artifact which contains it. The
core library can't take dependencies on the management artifacts,
and we had to do some nasty workaround to add a dependency to the
tools artifact. Moving EspAuthenticator outside of the core JAR
also takes it off of the system classpath for tools, which means
it should run without any workaround. Frameworks users will depend
on this JAR rather than endpoints-management-auth.

This is added to the management repository to share version numbers
with the management artifacts for consistency.

Change-Id: I0c6fa4264054743fef1c481b0ef8f3cf64ef7881