only send consumer metrics when API key is valid
This prevents the API producer project from seeing requests not tied to
any particular consumer from appeaering in its consumer graphs.
Previously, a request using an API key from the producer project (to its
own api) and a request with no API key would both appear in the consumer
graphs, which is misleading.

Change-Id: Iba9db04fc38e83970a191c5c89f61b4388d19a0a