fix default version config fetching
A previous change changed the URL template for fetching the service
config when a version is not specified. However, it was faulty. This
change addresses the issue by importing the Service Management client
library, which is used over hardcoded URL templates. When a version is
not specified, the list of config versions is fetched and the first
config id is used for fetching the individual configuration.