add better error handling for config filter init
Right now, if the service config fetching fails, it crashes the
instance due to throwing a RuntimeException, while the filter
initialization catches only IOException. This change will also catch
ServiceConfigException.

Current ESP behavior is to crash the instance if service config
fetching fails; however, this behavior is not as desirable, especially
from App Engine, which will aggressively and quickly retry
initialization, which can potentially exhaust quota. Thus, this change
will log the error, but allow initialization to continue, but deny all
requests that pass through the filter to ensure that the API is not
unduly accessible.

Lastly, this change adds some extensive help text to a few specific
error codes that the Service Management API may return, to help user
debugging.