Increase the version number for the development of the next release. (#54)
* Increase the version number for the development of the next release.

* Set Travis CI Java version to openjdk8

This framework is intended to work with Java 8 on Google App Engine
Standard. Newer versions are not supported.

* Make platform discovery hermetic in unit tests.

This fixes failing Travis CI tests. Now Travis moved to GCE and that
environment is reflected in ReportRequest labels in ControlFilter tests.
Unit tests shouldn't probe the execution environment, so a hermetic HTTP
transport to check for "Metadata Server" existence is passed to the
ControlFilter constructor.

* Set version to 1.0.12 (drop -SNAPSHOT)