Remove service serialization to debug log in ConfigFilter
The ConfigFilter used to print a whole string serialization of the
corresponding service object to the debug log. As the service object can
become very large, this consumes many CPU cycles even when the debug log
is not enabled.

As the log message with thousands of rows is not readable anyway, only the
service's name will be printed to the log now.