Bugfix: data race in Service Control Client package (#64)
* Avoid data races leading to uninformative log entries.

A data race between flushAndSchedule* methods and stop() causes a
NullPointerException. The exception itself does not affect the operation
of the library, but generates misleading log entries.

Context:
Under normal condiions flushAndSchedule* methods repeatedly schedule
their own execution at constant time intervals. This way the library
achieves regular execution of flushes. The execution should stop when
the client is stopped. To achieve this, flushAndSchedule* methods reset
counters and return if they detect a stopped client.

Race condition:
Client.stop() method might be called in one thread, while
flushAndSchedule* are active. The method stop resets the field scheduler
which in a while is dereferenced at the end of flushAndSchedule* methods.

Solution:
In flushAndSchedule* methods, check if the client is stopped for the
second time towards the end of the function, just before dereferencing
the field scheduler. Also make a local copy of the field scheduler.

* Bump version to 1.0.14